solucionFormulario

+++ products/forms.py
@@ -2,4 +2,4 @@ from django import forms
 
 class Formulario(forms.Form):
     refresco = forms.CharField(max_length=40)
-    cont = forms.FloatField()
+    cont = forms.IntegerField()
